+++ src/quacc/schemas/ase.py
@@ -11,8 +11,8 @@ from ase.io import read
 from ase.vibrations.data import VibrationsData
 
 from quacc import SETTINGS
-from quacc.runners.prep import prep_next_run as prep_next_run_
 from quacc.schemas.atoms import atoms_to_metadata
+from quacc.schemas.prep import prep_next_run as prep_next_run_
 from quacc.utils.dicts import recursive_merge_dicts, sort_dict
 from quacc.utils.files import get_uri
 from quacc.wflow.db import results_to_db